+++ src/Data/Singletons/TypeLits.hs
@@ -1,3 +1,6 @@
+{-# LANGUAGE TemplateHaskell, ScopedTypeVariables, TypeInType, ConstraintKinds,
+             GADTs, TypeFamilies #-}
+
 -----------------------------------------------------------------------------
 -- |
 -- Module      :  Data.Singletons.TypeLits
@@ -18,13 +21,16 @@ module Data.Singletons.TypeLits (
   Sing(SNat, SSym),
   SNat, SSymbol, withKnownNat, withKnownSymbol,
   Error, ErrorSym0, ErrorSym1, sError,
-  KnownNat, natVal, KnownSymbol, symbolVal,
+  KnownNat, KnownNatSym0, KnownNatSym1, natVal,
+  KnownSymbol, KnownSymbolSym0, KnownSymbolSym1, symbolVal,
 
   (:^), (:^$), (:^$$), (:^$$$)
   ) where
 
 import Data.Singletons.TypeLits.Internal
 import Data.Singletons.Prelude.Num ()   -- for typelits instances
+
+import Data.Singletons.Promote
 
 -- This bogus Num instance is helpful for people who want to define
 -- functions over Nats that will only be used at the type level or
@@ -40,3 +46,6 @@ instance Num Nat where
 
 no_term_level_nats :: a
 no_term_level_nats = error "The kind `Nat` may not be used at the term level."
+
+-- These are often useful in TypeLits-heavy code
+$(genDefunSymbols [''KnownNat, ''KnownSymbol])